Fixed DataLoader, Fixed Scrolling list in Library window
Added WindowEvents which just defines a method onWindowShown for
LibraryEvents. I might automate the calling of that function, but as of
now, it has to be hard coded into the Driver. Additionally, I added
GuiObject.SetParent() which allows you to set a parent to another object
in one move without getting the parent and then adding a child. It also
handles the case when the new or old parent is to be null.
I have yet to define how far the user should be able to scroll.