Polishing edits.
Fixed up gui in management tab for adding a new items. Removed the "edit
items" button because it was really glitchy and was resulting in certain
albums randomly disappearing at times. I replaced it with an edit button
on the library page for administrators. Got a "Play" button on the
library page working.